+++ account_balance_line/README.rst
@@ -6,7 +6,7 @@
 Account Balance Line
 ====================
 
-Add a balance total for grouped lines in move line view.
+This module adds a balance total for grouped lines in move line view.
 
 Balance field will only be shown when move lines are grouped by account
 or filtered by account.
@@ -14,21 +14,23 @@ or filtered by account.
 Usage
 =====
 
+#. If you're an accounting adviser, enable developer mode and go to
+   Accounting > Adviser > Journal items.
+#. Or open an account or a partner, and click on Action > Journal items.
+#. Filter or group by account.
+#. You will see a column named "Balance" with the balance of the line.
+
 .. image:: https://odoo-community.org/website/image/ir.attachment/5784_f2813bd/datas
    :alt: Try me on Runbot
-   :target: https://runbot.odoo-community.org/runbot/92/8.0
-
-For further information, please visit:
-
- * https://www.odoo.com/forum/help-1
+   :target: https://runbot.odoo-community.org/runbot/92/9.0
 
 Bug Tracker
 ===========
 
-Bugs are tracked on `GitHub Issues <https://github.com/OCA/account-financial-tools/issues>`_.
-In case of trouble, please check there if your issue has already been reported.
-If you spotted it first, help us smashing it by providing a detailed and welcomed feedback
-`here <https://github.com/OCA/account-financial-tools/issues/new?body=module:%20account_balance_line%0Aversion:%208.0%0A%0A**Steps%20to%20reproduce**%0A-%20...%0A%0A**Current%20behavior**%0A%0A**Expected%20behavior**>`_.
+Bugs are tracked on `GitHub Issues
+<https://github.com/OCA/account-financial-tools/issues>`_. In case of trouble,
+please check there if your issue has already been reported. If you spotted it
+first, help us smashing it by providing a detailed and welcomed feedback.
 
 Credits
 =======
@@ -38,6 +40,7 @@ Contributors
 
 * Vincent revaville <vincent.renaville@camptocamp.com>
 * Yannick Vaucher <yannick.vaucher@camptocamp.com>
+* Vicent Cubells <vicent.cubells@tecnativa.com>
 
 Maintainer
 ----------
