+++ account_balance_line/__manifest__.py
@@ -1,34 +1,23 @@
 # -*- coding: utf-8 -*-
-##############################################################################
-#
-#    Author: Vincent Renaville (Camptocamp)
-#    Copyright 2010-2014 Camptocamp SA
-#
-#    This program is free software: you can redistribute it and/or modify
-#    it under the terms of the GNU Affero General Public License as
-#    published by the Free Software Foundation, either version 3 of the
-#    License, or (at your option) any later version.
-#
-#    This program is distributed in the hope that it will be useful,
-#    but WITHOUT ANY WARRANTY; without even the implied warranty of
-#    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-#    GNU Affero General Public License for more details.
-#
-#    You should have received a copy of the GNU Affero General Public License
-#    along with this program.  If not, see <http://www.gnu.org/licenses/>.
-#
-##############################################################################
-
-{'name': 'Balance on lines',
- 'summary': 'Display balance totals in move line view',
- 'version': '8.0.1.1.1',
- 'author': "Camptocamp,Odoo Community Association (OCA)",
- 'maintainter': 'Camptocamp',
- 'category': 'Accounting',
- 'depends': ['account'],
- 'website': 'http://www.camptocamp.com',
- 'data': ['account_move_line_view.xml'],
- 'tests': [],
- 'installable': True,
- 'license': 'AGPL-3',
- }
+# Copyright 2010-2014 Camptocamp - Vincent Renaville
+# Copyright 2016 Tecnativa - Vicent Cubells
+# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
+{
+    'name': 'Balance on lines',
+    'summary': 'Display balance totals in move line view',
+    'version': '9.0.1.0.0',
+    'category': 'Accounting',
+    'website': 'http://www.tecnativa.com,',
+    'author': "Camptocamp,"
+              "Tecnativa,"
+              "Odoo Community Association (OCA)",
+    'license': 'AGPL-3',
+    'application': False,
+    'installable': True,
+    'depends': [
+        'account'
+    ],
+    'data': [
+        'views/account_move_line_view.xml'
+    ],
+}
