--- account_balance_line/account_move_line.py
@@ -1,37 +1,0 @@
-# -*- coding: utf-8 -*-
-##############################################################################
-#
-#    Author: Vincent Renaville (Camptocamp)
-#    Copyright 2010-2014 Camptocamp SA
-#
-#    This program is free software: you can redistribute it and/or modify
-#    it under the terms of the GNU Affero General Public License as
-#    published by the Free Software Foundation, either version 3 of the
-#    License, or (at your option) any later version.
-#
-#    This program is distributed in the hope that it will be useful,
-#    but WITHOUT ANY WARRANTY; without even the implied warranty of
-#    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-#    GNU Affero General Public License for more details.
-#
-#    You should have received a copy of the GNU Affero General Public License
-#    along with this program.  If not, see <http://www.gnu.org/licenses/>.
-#
-##############################################################################
-
-from openerp import models, fields, api
-import openerp.addons.decimal_precision as dp
-
-
-class AccountMoveLine(models.Model):
-    _inherit = "account.move.line"
-
-    @api.multi
-    @api.depends('credit', 'debit')
-    def _line_balance(self):
-        for line in self:
-            line.line_balance = line.debit - line.credit
-
-    line_balance = fields.Float(
-        compute='_line_balance', string='Balance', store=True,
-        digits=dp.get_precision('Account'))
