--- account_balance_line/account_move_line_view.xml
@@ -1,30 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<openerp>
-  <data>
-
-    <record id="view_account_move_line_filter_balance" model="ir.ui.view">
-      <field name="name">Journal Items add visibilty for balance</field>
-      <field name="model">account.move.line</field>
-      <field name="inherit_id" ref="account.view_account_move_line_filter"/>
-      <field name="arch" type="xml">
-        <field name="account_id" position="attributes">
-          <attribute name="context">{'invisible_balance': False}</attribute>
-        </field>
-        <filter string="Account" position="attributes">
-          <attribute name="context">{'group_by': 'account_id', 'invisible_balance': False}</attribute>
-        </filter>
-      </field>
-    </record>
-
-    <record model="ir.ui.view" id="account_move_line_balance_custom">
-      <field name="name">account_move_line_balance_custom</field>
-      <field name="model">account.move.line</field>
-      <field name="inherit_id" ref="account.view_move_line_tree"/>
-      <field name="arch" type="xml">
-        <field name="credit" position="after" >
-          <field name="line_balance" sum="Total Balance" invisible="context.get('invisible_balance', True)"/>
-        </field>
-      </field>
-    </record>
-  </data>
-</openerp>
