--- account_balance_line/i18n/account_balance_line.pot
@@ -1,53 +1,0 @@
-# Translation of OpenERP Server.
-# This file contains the translation of the following modules:
-#	* account_balance_line
-#
-msgid ""
-msgstr ""
-"Project-Id-Version: OpenERP Server 7.0\n"
-"Report-Msgid-Bugs-To: \n"
-"POT-Creation-Date: 2014-06-20 14:11+0000\n"
-"PO-Revision-Date: 2014-06-20 14:11+0000\n"
-"Last-Translator: <>\n"
-"Language-Team: \n"
-"MIME-Version: 1.0\n"
-"Content-Type: text/plain; charset=UTF-8\n"
-"Content-Transfer-Encoding: \n"
-"Plural-Forms: \n"
-
-#. module: account_balance_line
-#: view:account.move.line:0
-msgid "Account"
-msgstr ""
-
-#. module: account_balance_line
-#: model:ir.model,name:account_balance_line.model_account_move_line
-msgid "Entry lines"
-msgstr ""
-
-#. module: account_balance_line
-#: code:_description:0
-#, python-format
-msgid "Journal Items"
-msgstr ""
-
-#. module: account_balance_line
-#: field:account.move.line,line_balance:0
-msgid "Balance"
-msgstr ""
-
-#. module: account_balance_line
-#: view:account.move.line:0
-msgid "Total Balance"
-msgstr ""
-
-#. module: account_balance_line
-#: view:account.move.line:0
-msgid "{'group_by': 'account_id', 'invisible_balance': False}"
-msgstr ""
-
-#. module: account_balance_line
-#: view:account.move.line:0
-msgid "{'invisible_balance': False}"
-msgstr ""
-
