+++ account_balance_line/i18n/es.po
@@ -1,49 +1,32 @@
 # Translation of Odoo Server.
 # This file contains the translation of the following modules:
-# * account_balance_line
-# 
-# Translators:
-# Carles Antolí <carlesantoli@hotmail.com>, 2015
+#	* account_balance_line
+#
 msgid ""
 msgstr ""
-"Project-Id-Version: account-financial-tools (8.0)\n"
+"Project-Id-Version: Odoo Server 9.0c\n"
 "Report-Msgid-Bugs-To: \n"
-"POT-Creation-Date: 2015-12-05 01:25+0000\n"
-"PO-Revision-Date: 2015-12-06 18:21+0000\n"
-"Last-Translator: Carles Antolí <carlesantoli@hotmail.com>\n"
-"Language-Team: Spanish (http://www.transifex.com/oca/OCA-account-financial-tools-8-0/language/es/)\n"
+"POT-Creation-Date: 2016-12-17 12:16+0000\n"
+"PO-Revision-Date: 2016-12-17 12:16+0000\n"
+"Last-Translator: <>\n"
+"Language-Team: \n"
 "MIME-Version: 1.0\n"
 "Content-Type: text/plain; charset=UTF-8\n"
 "Content-Transfer-Encoding: \n"
-"Language: es\n"
-"Plural-Forms: nplurals=2; plural=(n != 1);\n"
+"Plural-Forms: \n"
 
 #. module: account_balance_line
-#: view:account.move.line:account_balance_line.view_account_move_line_filter_balance
-msgid "Account"
-msgstr "Cuenta"
-
-#. module: account_balance_line
-#: field:account.move.line,line_balance:0
+#: model:ir.model.fields,field_description:account_balance_line.field_account_move_line_line_balance
 msgid "Balance"
 msgstr "Saldo"
 
 #. module: account_balance_line
 #: model:ir.model,name:account_balance_line.model_account_move_line
-msgid "Journal Items"
-msgstr "Apuntes contables"
+msgid "Journal Item"
+msgstr "Apunte contable"
 
 #. module: account_balance_line
-#: view:account.move.line:account_balance_line.account_move_line_balance_custom
+#: model:ir.ui.view,arch_db:account_balance_line.account_move_line_balance_custom
 msgid "Total Balance"
 msgstr "Total saldo"
 
-#. module: account_balance_line
-#: view:account.move.line:account_balance_line.view_account_move_line_filter_balance
-msgid "{'group_by': 'account_id', 'invisible_balance': False}"
-msgstr "{'group_by': 'account_id', 'invisible_balance': False}"
-
-#. module: account_balance_line
-#: view:account.move.line:account_balance_line.view_account_move_line_filter_balance
-msgid "{'invisible_balance': False}"
-msgstr "{'invisible_balance': False}"
