+++ account_balance_line/models/__init__.py
@@ -3,4 +3,4 @@
 # Copyright 2016 Tecnativa - Vicent Cubells
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
 
-from . import models
+from . import account_move_line
